% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalDistributions.R
\name{compute_A_matrix}
\alias{compute_A_matrix}
\title{Compute conditional distribution matrix}
\usage{
compute_A_matrix(R_bins, cond, numeric.points = 3, verbose = F)
}
\arguments{
\item{R_bins}{The number of latent bins to choose}

\item{cond}{Conditional distribution}

\item{verbose}{Verbose parameter}
}
\value{
Conditional Distribution Matric Form
}
\description{
Compute conditional distribution matrix
}
