% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicMath.R
\name{compute_edf}
\alias{compute_edf}
\title{Compute Empirical Distribution Function}
\usage{
compute_edf(x, N, weights)
}
\arguments{
\item{x}{sample of observations; vector of integers}

\item{N}{(support size {0,...,N}); integer}
}
\value{
empirical distribution over integers {0,...,N}
}
\description{
Compute Empirical Distribution Function
}
