% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalDistributions.R
\name{conditional_mkm}
\alias{conditional_mkm}
\title{Conditional distribution function for the measurement kernel model}
\usage{
conditional_mkm(N, ker, h)
}
\arguments{
\item{N}{Number of Questions}

\item{ker}{Measurement Kernel}

\item{h}{Bandwidth}
}
\value{
Conditional distribution function for the measurement kernel model
}
\description{
Conditional distribution function for the measurement kernel model
}
