% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVparams.R
\name{cv_regression_gridsearch}
\alias{cv_regression_gridsearch}
\title{Cross validation for regression and regularization parameters}
\usage{
cv_regression_gridsearch(
  data,
  cond,
  folds = 5,
  outcome = 1,
  X.cols = seq(2, ncol(data)),
  ker.params,
  reg.kers,
  mu.params,
  threshold = 10^(-5),
  max.iter = 50
)
}
\arguments{
\item{data}{Data used to estimate the model}

\item{cond}{Conditional Model}

\item{folds}{Number of folds}

\item{outcome}{Which column is the outcome}

\item{X.cols}{Which columns are covariates used}

\item{ker.params}{Parameters to input to scale each parameter}

\item{reg.kers}{Kernels used for smoothing across covariates}

\item{mu.params}{Parameter vector for the regularization term}

\item{threshold}{Threshold for EM algorithm likelihood change}

\item{max.iter}{Maximum number of iterations for EM algorithm}
}
\value{
List of cross validation likelihood, and optimal parameters
}
\description{
Cross validation for regression and regularization parameters
}
