% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConversionRegression.R
\name{empirical_cross_entropy}
\alias{empirical_cross_entropy}
\title{Evaluation of Average Cross Entropy On A Test Set}
\usage{
empirical_cross_entropy(p.mat, z.vec)
}
\arguments{
\item{p.mat}{Matrix of predicted probabilities}

\item{z.vec}{Vector of outcomes (Train or test)}
}
\value{
Normalized Cross Entropy
}
\description{
Evaluation of Average Cross Entropy On A Test Set
}
