% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDistributions.R
\name{estimate_mixing_npem}
\alias{estimate_mixing_npem}
\title{Nonparametric EM Algorithm for latent trait}
\usage{
estimate_mixing_npem(
  p.hat,
  A.matrix,
  mu = 0,
  threshold = 5 * 10^(-5),
  verbose = F,
  max.iter = 50,
  init.latent
)
}
\arguments{
\item{p.hat}{Discrete empirical distribution function}

\item{A.matrix}{Conditional distribution Matrix}

\item{mu}{Regularization Parameter}

\item{threshold}{Threshold for the change in the likelihood}

\item{verbose}{Whether to display progress}

\item{max.iter}{Maximum number of iterations of the EM algorithm}

\item{init.latent}{Initialize latent distribution.  This must be}
}
\value{
latent trait and observed distribution
}
\description{
Nonparametric EM Algorithm for latent trait
}
