% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDistributions.R
\name{estimate_mixing_numeric}
\alias{estimate_mixing_numeric}
\title{CVX implementation}
\usage{
estimate_mixing_numeric(p.hat, A.matrix, mu = 0, cvx.solver = "SCS")
}
\arguments{
\item{p.hat}{Discrete empirical distribution function}

\item{A.matrix}{Conditional distribution Matrix}

\item{mu}{Regularization Parameter}

\item{cvx.solver}{Which solver to use}
}
\value{
latent trait and observed distribution
}
\description{
CVX implementation
}
