% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalDistributions.R
\name{generate_cond_binomial}
\alias{generate_cond_binomial}
\title{Conditional distribution function for the binomial model}
\usage{
generate_cond_binomial(N)
}
\arguments{
\item{N}{Number of Questions}
}
\value{
Conditional distribution function for the binomial model
}
\description{
Conditional distribution function for the binomial model
}
