% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{generate_mkm_list}
\alias{generate_mkm_list}
\title{Generate List of Measurement Kernel Models}
\usage{
generate_mkm_list(N, ker, h.set)
}
\arguments{
\item{N}{Number of questions}

\item{ker}{Kernel function}

\item{h.set}{Vector of bandwidths}
}
\value{
List of kernel functions
}
\description{
Generate List of Measurement Kernel Models
}
