% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConversionRegression.R
\name{ImputationRegressionGEE}
\alias{ImputationRegressionGEE}
\title{GEE Regression from imputed data}
\usage{
ImputationRegressionGEE(formula, X, Z.impute, ...)
}
\arguments{
\item{formula}{regression formula}

\item{X}{covariates}

\item{Z.impute}{imputed outcomes}
}
\description{
GEE Regression from imputed data
}
