% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConversionRegression.R
\name{ImputeOutcomes}
\alias{ImputeOutcomes}
\title{Impute Outcomes}
\usage{
ImputeOutcomes(
  X.ref,
  y.ref,
  z.ref,
  n.impute,
  Y.train,
  Z.train,
  cond.y,
  cond.z,
  mu.y,
  mu.z,
  ref.cols,
  ker.set,
  R.bins = 1000,
  threshold = 5 * 10^(-5)
)
}
\arguments{
\item{X.ref}{Covariates}

\item{y.ref}{Outcome 1}

\item{z.ref}{Outcome 2 (imputed outcome)}

\item{n.impute}{Number of times to impute}

\item{Y.train}{Training data}

\item{Z.train}{Testing data}

\item{cond.y}{Conditional distribution 1}

\item{cond.z}{Conditional distribution 2}

\item{mu.y}{Tuning parameter 1}

\item{mu.z}{Tuning parameter 2}

\item{ref.cols}{reference columns}

\item{ker.set}{Kernels for smoothing over X parameters}

\item{R.bins}{Number of bins for model parameter}
}
\description{
Impute Outcomes
}
