% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConversionRegression.R
\name{normalScoreConversionProb}
\alias{normalScoreConversionProb}
\title{Naive Conversion Method}
\usage{
normalScoreConversionProb(y.vec, muy, sdy, muz, sdz, Nz, n.samp = 20000)
}
\description{
Naive Conversion Method
}
