% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConversionRegression.R
\name{predictedDistributions}
\alias{predictedDistributions}
\title{Predict Distributions}
\usage{
predictedDistributions(
  X.ref,
  y.ref,
  Y.train,
  Z.train,
  cond.y,
  cond.z,
  mu.y,
  mu.z,
  ref.cols,
  ker.set,
  R.bins = 1000,
  threshold = 5 * 10^(-5),
  max.iter = 50,
  verbose = F
)
}
\arguments{
\item{X.ref}{Covariates Reference}

\item{y.ref}{Observed Data Reference (to convert)}

\item{Y.train}{Training data}

\item{Z.train}{Testing data}

\item{cond.y}{Conditional distribution 1}

\item{cond.z}{Conditional distribution 2}

\item{mu.y}{Tuning parameter 1}

\item{mu.z}{Tuning parameter 2}

\item{ref.cols}{reference columns}

\item{ker.set}{Kernels for smoothing over X parameters}

\item{R.bins}{Number of bins for model parameter}

\item{threshold}{Threshold for NPEM Iterations}

\item{max.iter}{Maximum number of iterations in the NPEM algorithm}

\item{verbose}{Print fitting details}
}
\description{
Predict Distributions
}
