% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{second_order_feasibility_test}
\alias{second_order_feasibility_test}
\title{Second Order Feasibility Test}
\usage{
second_order_feasibility_test(latent.mixture, A.tensor, p.hat, sample.size)
}
\arguments{
\item{latent.mixture}{Latent Distribution}

\item{A.tensor}{Conditional distribution tensor}

\item{p.hat}{Empirical Distribution}

\item{sample.size}{Sample Size}
}
\value{
p value for second order feasibility test
}
\description{
Second Order Feasibility Test
}
