% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentDistributions.R
\name{select_R}
\alias{select_R}
\title{Select Number of Bins}
\usage{
select_R(
  p.hat,
  cond,
  mu = 0.001,
  R.min = 100,
  R.max = 5000,
  threshold = 10^(-4),
  num.steps = 20,
  max.iter = 50
)
}
\arguments{
\item{p.hat}{Distribution to approximate}

\item{cond}{Conditional distribution (Empirical Bayes f model)}

\item{mu}{Regularization Parameter}

\item{R.min}{Minimum number of bins}

\item{R.max}{Maximum number of bins}

\item{threshold}{Threshold for log-likelihood difference}

\item{num.steps}{Number of steps to take for the R (number of bins) selection}

\item{max.iter}{Maximum iterations of the NPEM algorithm.}
}
\value{
Smallest viable value of R
}
\description{
Select Number of Bins
}
