% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{first_order_feasibility_test}
\alias{first_order_feasibility_test}
\title{First order feasibility test}
\usage{
first_order_feasibility_test(latent.mixture, A.matrix, p.hat, sample.size)
}
\arguments{
\item{p.hat}{Empirical Distribution}

\item{sample.size}{Sample size}

\item{p.ma}{Nonparametric Maximum Likelihood Marginal Distribution}
}
\value{
p value for first order feasibility test
}
\description{
First order feasibility test
}
